Move basal body per-cell color to config file.
Before, the program would crash if there was
more than the number of hard-coded cell colors.

Also, Ximena thought there should be an option
for make all basal bodys in cells the same color.

But it's still handy for debugging to use different
colors for different cells.

Now, in the config file, it is possible to set
the default and not-in-any-cell basal body colors.

It's also possible, but not required, to give
a number of different colors, each assigned
to a different cell.  If the number of cells
exceeds the number of colors, the default
color gets used for the extra cells.