Get rid of DISPLAY os.environ code
It was causing a crash on MS-Windows.

I used to use it for working on the program using ssh.
matplotlib needs the DISPLAY to be set, which it isn't
over ssh.  But I could still test the program up to
the creation of the graphs.

Maybe in the future we'll bring the capability back.