Add image without cell vertices center mark
The image that shows the center of each cell shows two centers.
 * x marks the center of cell vertices
 * o marks the center of the cell plane figure

The center of vertices is easy to calculate.  It's needed to
inspect how the vertices relate to the center calculations.
But it has no significance outside from verifying that the
calculations are valid.

The center of the plane figure is the important one.  It is
calculated using the center of vertices.

So we want an image showing both so that we can inspect our work.
That is what we already had before this commit (fname_z4).

But to show others, we want an image showing just the plane figure
cell centers.  That's the new image in this commit (fname_z5).